Typescript version: Publish ES modules
This allows applications to decide to transpile and bundle instead.
This also allows libraries that depend on this to be published
as ES modules without bundling themselves. Existing tools (webpack,
Parcel, Rollup) do well at bundling dependencies authored in ES
modules, so this seems like a better default. It might be wrong
though!